Move fusiondoc model into the `runtime` package.
