+++ fusioncli/src/main/java/dev/ionfusion/fusioncli/doc/tool/ExportedBinding.java
@@ -5,7 +5,7 @@ package dev.ionfusion.fusioncli.doc.tool;
 
 import static java.util.Comparator.comparing;
 
-import dev.ionfusion.fusion._private.doc.model.BindingDoc;
+import dev.ionfusion.runtime._private.doc.BindingDoc;
 import dev.ionfusion.runtime.base.ModuleIdentity;
 import java.util.Comparator;
 import java.util.List;
