+++ fusioncli/src/main/java/dev/ionfusion/fusioncli/doc/tool/RepoEntity.java
@@ -7,7 +7,7 @@ import static dev.ionfusion.fusion._Private_Trampoline.discoverModulesInReposito
 import static dev.ionfusion.fusion._Private_Trampoline.instantiateModuleDocs;
 import static dev.ionfusion.fusion._Private_Trampoline.loadModule;
 
-import dev.ionfusion.fusion._private.doc.model.ModuleDocs;
+import dev.ionfusion.runtime._private.doc.ModuleDocs;
 import dev.ionfusion.runtime.base.FusionException;
 import dev.ionfusion.runtime.base.ModuleIdentity;
 import dev.ionfusion.runtime.embed.TopLevel;
