+++ runtime/src/main/java/dev/ionfusion/fusion/ModuleBuilderImpl.java
@@ -7,7 +7,7 @@ import com.amazon.ion.IonString;
 import com.amazon.ion.IonStruct;
 import com.amazon.ion.IonValue;
 import dev.ionfusion.fusion.Namespace.NsDefinedBinding;
-import dev.ionfusion.fusion._private.doc.model.BindingDoc.Kind;
+import dev.ionfusion.runtime._private.doc.BindingDoc.Kind;
 import dev.ionfusion.runtime.base.FusionException;
 import dev.ionfusion.runtime.base.ModuleIdentity;
 import dev.ionfusion.runtime.embed.ModuleBuilder;
