+++ runtime/src/main/java/dev/ionfusion/fusion/ModuleInstance.java
@@ -8,10 +8,11 @@ import dev.ionfusion.fusion.ModuleNamespace.DefinedProvidedBinding;
 import dev.ionfusion.fusion.ModuleNamespace.ModuleDefinedBinding;
 import dev.ionfusion.fusion.ModuleNamespace.ProvidedBinding;
 import dev.ionfusion.fusion.Namespace.NsDefinedBinding;
-import dev.ionfusion.fusion._private.doc.model.BindingDoc;
-import dev.ionfusion.fusion._private.doc.model.ModuleDocs;
+import dev.ionfusion.runtime._private.doc.BindingDoc;
+import dev.ionfusion.runtime._private.doc.ModuleDocs;
 import dev.ionfusion.runtime.base.FusionException;
 import dev.ionfusion.runtime.base.ModuleIdentity;
+import dev.ionfusion.runtime.embed.ModuleBuilder;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
