+++ runtime/src/main/java/dev/ionfusion/fusion/_Private_Trampoline.java
@@ -3,7 +3,7 @@
 
 package dev.ionfusion.fusion;
 
-import dev.ionfusion.fusion._private.doc.model.ModuleDocs;
+import dev.ionfusion.runtime._private.doc.ModuleDocs;
 import dev.ionfusion.runtime.base.FusionException;
 import dev.ionfusion.runtime.base.ModuleIdentity;
 import dev.ionfusion.runtime.embed.TopLevel;
