+++ runtime/src/main/java/dev/ionfusion/runtime/_private/doc/BindingDoc.java
@@ -1,7 +1,7 @@
 // Copyright Ion Fusion contributors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package dev.ionfusion.fusion._private.doc.model;
+package dev.ionfusion.runtime._private.doc;
 
 public final class BindingDoc
 {
